function in java 8

+++ demo/src/com/demo/java8/Employee.java
@@ -4,10 +4,10 @@ public class Employee {
 
 	int id;
 	String name;
-	double salary;
+	int salary;
 
 	
-	public Employee(int id, String name, double salary) {
+	public Employee(int id, String name, int salary) {
 		super();
 		this.id = id;
 		this.name = name;
